Fixes gulp-changed plugin's usage in the gulpfile.js
Fixes wrong arguments in $.changed() which should be passing only changed .scss files
for complilation to .css.
First argument was 'styles' which is a non-existant folder since the destination forlder
is .tmp/styles. The second argument should have the suffices of the destination folder files,
which is .css and not .scss