Only report HTML size in HTML task
In the previous configuration size was also reporting the size of JavaScript,
which usref adds to the stream. Open to other ways of filtering by type or
drying up this code, but didn't want to add a new (more complicated) pattern to
WSK with lazypipe or pumpify to save complex streams.